Restored extra dir level for universal includes
This extra level is necessary to do linked inclusion. Without the extra
level of nesting, including the root include directory will build all
msp device libraries (which will cause build errors). When linking to an
include directory, the link should point to the device-specific
directory (which should contain a symbolic link to the generic msp
includes). Silly issue that I've resolved in the past and re-created by
restructuring the directories.

Also added links for each currently supported part and corrected some
include files which pointed to the old path.