g2553/spi: Added G2553 SPI library
This library provides all of the basic functionality needed to get SPI
up and running on an MSP430G2553 device. Example with logic analyzer
output example provided as well. Keep in mind this library only provides
SPI output functionality for now. If we come up with a good way to
support universal SPI input, we will add that in the future.