*/pin_fw: Improved perf & moved data to flash
After realizing that I must have been puffing something to write code
that has division and modulus operations for something so consistently
used, it was time to fix it. We almost gave up on our silly little pin
framework, but instead let's give it another go. I'd done this design in
the first place, but this time I've taken care to put all of the data in
flash memory rather than RAM. I think this is as optimized as the
performance will get with any type of lookup. We can make the data more
efficient, but it will still be at the cost of additional overhead in
some form or fashion. Anyone with tips on how this can be improved is
more than welcome to provide them.

Also fixed SPI libraries which try to define a const array (it's already
const now, so can't do that...).

Enjoy.