f5510/isr_dispatcher: Add the ISR dispatcher
Add a new module called the ISR dispatcher. This module implements an
ISR for each interrupt vector and allows runtime installation and
uninstallation of functions to be called from the respective ISR.

Adding this functionality allows more effective dynamic allocation of
various components within the device. Need to also add lookup capability
for interrupts on both pins and timers.