f5510/clock & isr_d: Add ISR dispatcher in clock
Add ISR dispatcher installation into the clock library. Also add
functionality to the ISR dispatcher to allow installed ISR functions to
modify the SR on exit as they would be able to do if installed natively.

This unfortunately adds overhead in terms of latency and code size, but
without this capability there is no way to allow, for example, the delay
ISR to exit the system from LPM. Compilation flags are also available to
disable the SR checking if desired, but this was left as a global
configuration for now. Might be a good idea to modularize it (as only
specific ISRs will need to modify the SR). Until that point, leave it
as-is and think of ways to optimize the SR checking.