f5510/spi: Make SOMI initialization optional
This can cause issues when an output only SPI instance is needed and the
SOMI pin is intended to be used for something else. Thus, we'll assume
the SPI instance is output only by default, and if it's specified to be
I/O, the SOMI pin will also get initialized.