msp/int_sw: Add interrupt-based switch library
This library takes a pin and callback function pointer in its
constructor. The callback function is executed when the switch event
triggers.

The constructor also allows the definition of edge type in the
constructor, but defaults to high-to-low detection. In this scenario,
the pin's pullup resistor is enabled and a high-to-low transition from
the switch being closed triggers the switch event. For the low-to-high
detection scenario, the pin's pulldown resistor is enabled and closing
the switch should result in a low-to-high transition. When this
transition is detected, the switch event triggers.

By allowing the allocation of a callback function, the interrupt-based
switch library is extremely dynamic and powerful. An example of the
usage of this library is also included with this commit.