msp/tlc5925: Add uint8_t overload for write
Since the inherited write functions take a uint8_t parameter, best to be
certain that the proper write function (16-bit) is called by adding
another overload. This is just an inline that calls the 16-bit write
function of the TLC. It's important to be aware of this so there aren't
any failed writes to the device, which could result in undefined
behavior.