g2553/basic_clock: Added cpp basic_clock library
This attempts to be complete with as minimal a footprint as possible.
The full clock library will support runtime delays and uptime like the
clock library for the f5510, but since those features might not always
be needed/used, it's better to have two separate libraries for clock
utilization with a basic library that leaves all these features out.
On a device with only 512B RAM, it's important to minimize footprint
wherever possible.